buggy output and input
Starting to see some actual results, the previous problem was with
buffering for msgpack not working correctly. Now some lines are drawn,
but since we don't handle line attributes or repetition, the output is
not correct.

Also hooked up on_u8 to key input, which works ok for the most part,
some keys (e.g. ESC) needs to be specially translated, and the non-u8
encoded inputs and mouse needs to be treated differently as well.

Rough estimate is that we are some 20% towards being usable for the
terminal emulator- level of integration, decent milestone before
tackling multigrid and popup.