title, paste from clipboard and highlighting fixes
There seem to be about two annoying bug left before we are at/beyond
terminal-emulator nvim and can go for the external-messages, menus,
multigrid and popups. After that there should be little left.

The first one is that visual select does not actually seem to have
a visual effect, so either we are transferring attributes incorrectly
or there is something more sinister afoot.

The second are rendering artifacts when using deletion, the easiest
way to reproduce is to write something like

    test.testsson

and dw on test.