this fixes the missing selection code
The cause was simply that the type-test for fg/bg filtered out all
the boolean attributes - missing the inverse attribute.

The other visual anomaly bug is more interesting - as it is two in
one, triggered through dw and undo in tandem.

With one bug, multiple dws seem to give the wrong/missing cursor
position until the next package (might be buffering) as we actually
don't receive the goto command. Such buffering issues may explain other
visible glitches.

The other actually sends the characters, but for some reason they
either don't get rendered or presented properly. First thing to rule
out is if it is front/back buffer issue in tui, or dirty rectangles
in the chain from tui / onwards. This can be done by sneaking in a
screenshot- trigger for the full raster stage.